asn1-parse: clarify usage of onNextContainer
The existing documentation of `Data.ASN1.Parse.onNextContainer` did not
make clear that the supplied function neads to handle all elements of
the container. This is due to the usage of `runParseASN1`, erroring on
non-complete consumption of state, which is only apparent when reading
the actual implementation code.

contributes to #37